fix(publishing): Publish the new site
This is a terrible way to fix the problem, but it works. Basically we
have two repositories outfittter-guide.github.io and
outfitter-guide-landing-page. The new landing page is in
outfitter-guide-landing-page which is set up to publish to a
subdirectory. This brings in the original index.html and hopefully makes
it so other stuff works. But, now we need to publish both sites to make
stuff work. Longer term, we need to fix outfitter-guide-landing-page to
publish to outfitter-guide.github.io and also make it so it's not
publishing to a subdirectory.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>